Move transform into another partial class. Because transform part might be larger after.
